dep! orm 3 & dbal 4

+++ config/database.php
@@ -2,7 +2,7 @@
 return [
 	'driver' => env('DB_DRIVER'),
 	'host' => env('DB_HOST'),
-	'port' => env('DB_PORT'),
+	'port' => (int)env('DB_PORT'),
 	'dbname' => env('DB_DATABASE'),
 	'prefix' => env('DB_TABLE_PREFIX'),
 	'user' => env('DB_USER'),
